+++ index.html
@@ -45,7 +45,7 @@
         </div>
       </header>
       <div class="container">
-        <ul class="tab">
+        <!-- <ul class="tab">
           <li>
             <a href="#">
               <div class="tab-item selected"><h3>상영 중</h3></div>
@@ -66,12 +66,13 @@
               <div class="tab-item"><h3>상영 예정</h3></div>
             </a>
           </li>
-        </ul>
+        </ul> -->
         <main>
           <section>
             <h2>지금 인기 있는 영화</h2>
             <ul class="thumbnail-list"></ul>
           </section>
+          <button class="more-button">더 보기</button>
         </main>
       </div>
